# Use byteToHex to decode uuid