--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -83,7 +83,7 @@ jobs:
           server-version: ${{ matrix.edgedb-version }}
 
       - name: Run query builder tests
-        if: ${{ matrix.node-version >= 14 && matrix.edgedb-version == 'nightly' }}  # change to stable once 2.0 lands
+        if: ${{ matrix.node-version >= 14 }}
         working-directory: qb
         run: |
           yarn test:ci