--- a/src/codecs/uuid.ts
+++ b/src/codecs/uuid.ts
@@ -35,11 +35,42 @@ function UUIDBufferFromString(uuid: string): Buffer {
   return buf;
 }
 
+const byteToHex: string[] = [];
+
+for (let i = 0; i < 256; ++i) {
+  byteToHex.push((i + 0x100).toString(16).slice(1));
+}
+export function bytesToHex(buffer: Buffer) {
+  const arr = new Uint8Array(buffer);
+
+  return (
+    byteToHex[arr[0]] +
+    byteToHex[arr[1]] +
+    byteToHex[arr[2]] +
+    byteToHex[arr[3]] +
+    "-" +
+    byteToHex[arr[4]] +
+    byteToHex[arr[5]] +
+    "-" +
+    byteToHex[arr[6]] +
+    byteToHex[arr[7]] +
+    "-" +
+    byteToHex[arr[8]] +
+    byteToHex[arr[9]] +
+    "-" +
+    byteToHex[arr[10]] +
+    byteToHex[arr[11]] +
+    byteToHex[arr[12]] +
+    byteToHex[arr[13]] +
+    byteToHex[arr[14]] +
+    byteToHex[arr[15]]
+  ).toLowerCase();
+}
+
 export class UUIDCodec extends ScalarCodec implements ICodec {
   encode(buf: WriteBuffer, object: any): void {
     if (typeof object === "string") {
-      const val = <string>object;
-      const ubuf = UUIDBufferFromString(val);
+      const ubuf = UUIDBufferFromString(object);
       buf.writeInt32(16);
       buf.writeBuffer(ubuf);
     } else {
@@ -50,6 +81,6 @@ export class UUIDCodec extends ScalarCodec implements ICodec {
   }
 
   decode(buf: ReadBuffer): any {
-    return buf.readBuffer(16).toString("hex");
+    return bytesToHex(buf.readBuffer(16));
   }
 }