Did some simple benchmarking on UUID parsing, see `qb/playground.ts` to see the various implementation.

I landed on an implementation from https://github.com/uuidjs/uuid/blob/main/src/stringify.js that's actually 22% faster than what we have currently and adds hyphens properly.

```
without-dashes: 821.39ms // current
with-dashes-bytetohex: 634.192ms // new
with-dashes-readbuffer: 1.714s
with-dashes-string-manip: 841.354ms
with-dashes-string-addition: 875.214ms
```